feat: Refactor slide deck system and introduce `SlideConfig`
*   Deprecates `Slide` in favor of a DSL-based `SlideConfig` and `DeckBuilder`.
*   Introduces `DeckNavigator` and `DeckHost` for simplified slide state management and transitions.
*   Moves legacy navigation and presentation components to a `deprecated` package.
*   Updates `SkiFrameFactory` to support opacity and refines default frame implementations.
