+++ composeApp/src/commonMain/kotlin/ke/don/ski/Deck.kt
@@ -5,11 +5,11 @@ import androidx.compose.runtime.remember
 import io.github.donald_okara.components.frames.defaultSkiFrames
 import io.github.donald_okara.components.values.Values
 import ke.don.domain.Slide
-import ke.don.ski.navigation.ContainerState
-import ke.don.ski.navigation.DeckNavigator
-import ke.don.ski.navigation.rememberContainerState
-import ke.don.ski.presentation.DeckMode
-import ke.don.ski.presentation.PresenterDsl
+import ke.don.ski.navigation.deprecated.ContainerState
+import ke.don.ski.navigation.deprecated.DeckNavigator
+import ke.don.ski.navigation.deprecated.rememberContainerState
+import ke.don.ski.presentation.deprecated.DeckMode
+import ke.don.ski.presentation.deprecated.PresenterDsl
 
 /**
  * Composable entry point that renders the slide deck UI with theme support, navigation, and frame layout.
@@ -28,8 +28,8 @@ fun Deck(
     },
     mode: DeckMode = DeckMode.Local
 ) {
-    val presentationFrame = defaultSkiFrames().snake.create(Values.cornerRadius)
-    val guidesFrame = defaultSkiFrames().basic.create(Values.cornerRadius)
+    val presentationFrame = defaultSkiFrames().snake.create(Values.cornerRadius, 0.5f)
+    val guidesFrame = defaultSkiFrames().basic.create(Values.cornerRadius, 0.5f)
 
     PresenterDsl(
         deckMode = mode,
