+++ composeApp/src/commonMain/kotlin/ke/don/ski/navigation/deprecated/ContainerState.kt
@@ -1,4 +1,4 @@
-package ke.don.ski.navigation
+package ke.don.ski.navigation.deprecated
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
