+++ composeApp/src/wasmJsMain/kotlin/DeckWebImpl.wasmJs.kt
@@ -1,13 +1,12 @@
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.ExperimentalComposeUiApi
 import ke.don.domain.Slide
 import ke.don.ski.Deck
-import ke.don.ski.navigation.DeckNavigator
-import ke.don.ski.navigation.DeckSyncState
-import ke.don.ski.navigation.rememberContainerState
-import ke.don.ski.presentation.DeckMode
+import ke.don.ski.navigation.deprecated.DeckNavigator
+import ke.don.ski.navigation.deprecated.DeckSyncState
+import ke.don.ski.navigation.deprecated.rememberContainerState
+import ke.don.ski.presentation.deprecated.DeckMode
 import kotlinx.browser.window
 import kotlinx.serialization.json.Json
 import org.w3c.dom.StorageEvent
